+++ models/WikiPageRevision.php
@@ -2,9 +2,10 @@
 
 namespace humhub\modules\wiki\models;
 
-use Yii;
 use humhub\components\ActiveRecord;
 use humhub\modules\user\models\User;
+use humhub\modules\wiki\permissions\CreatePage;
+use Yii;
 
 /**
  * This is the model class for table "wiki_page_revision".
@@ -16,6 +17,8 @@ use humhub\modules\user\models\User;
  * @property integer $wiki_page_id
  * @property integer $user_id
  * @property string $content
+ *
+ * @property-read WikiPage $page
  */
 class WikiPageRevision extends ActiveRecord
 {
@@ -73,6 +76,9 @@ class WikiPageRevision extends ActiveRecord
     public function beforeSave($insert)
     {
         $this->is_latest = 1;
+
+        $this->createPagesFromWikiLinks();
+
         return parent::beforeSave($insert);
     }
 
@@ -89,7 +95,74 @@ class WikiPageRevision extends ActiveRecord
             Yii::error($e);
         }
 
-        return parent::afterSave($insert, $changedAttributes);
+        parent::afterSave($insert, $changedAttributes);
+    }
+
+    /**
+     * Create Wiki Pages from wiki links found in the content of this Revision
+     */
+    private function createPagesFromWikiLinks()
+    {
+        if (!$this->page->content->container->can(CreatePage::class)) {
+            return;
+        }
+
+        if (!preg_match_all('/(\[.+?\]\(wiki:)(.+?)( ".+?"\))/i', $this->content, $wikiLinks)) {
+            return;
+        }
+
+        foreach ($wikiLinks[2] as $w => $wikiTitle) {
+            if ($wikiPage = $this->createOrFindPageByTitle($wikiTitle)) {
+                $this->updateWikiLink($wikiLinks[0][$w], $wikiLinks[1][$w] . $wikiPage->id . $wikiLinks[3][$w]);
+            }
+        }
+    }
+
+    /**
+     * Create or find existing Wiki Page by title
+     *
+     * @param string $title
+     * @return WikiPage|null
+     * @throws \yii\base\Exception
+     */
+    private function createOrFindPageByTitle(string $title): ?WikiPage
+    {
+        if (preg_match('/^\d+$/', $title) &&
+            WikiPage::find()->contentContainer($this->page->content->container)->where(['wiki_page.id' => $title])->exists()) {
+            // Exclude case when title is ID of the existing Wiki Page
+            return null;
+        }
+
+        if ($wikiPage = WikiPage::find()->contentContainer($this->page->content->container)->where(['title' => $title])->one()) {
+            return $wikiPage;
+        }
+
+        $wikiPage = new WikiPage($this->page->content->container);
+        $wikiPage->setScenario(WikiPage::SCENARIO_CREATE);
+        $wikiPage->title = $title;
+        $wikiPage->parent_page_id = $this->page->parent_page_id;
+        if (!$wikiPage->save()) {
+            return null;
+        }
+
+        $wikiPageRevision = $wikiPage->createRevision();
+        $wikiPageRevision->content = $title;
+        if (!$wikiPageRevision->save()) {
+            return null;
+        }
+
+        return $wikiPage;
+    }
+
+    /**
+     * Update Wiki Link in content
+     *
+     * @param string $wikiLinkWithTitle
+     * @param string $wikiLinkWithId
+     */
+    private function updateWikiLink(string $wikiLinkWithTitle, string $wikiLinkWithId)
+    {
+        $this->content = str_replace($wikiLinkWithTitle, $wikiLinkWithId, $this->content);
     }
 
 }