+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 - Enh #22: Diff view for Wiki page changes
 - Enh #138: Add "Save" action to sidebar navigation
 - Enh #9: Add activities for edited Wiki pages
+- Enh #140: Expand category hierarchy
 
 
 1.5.3  (April 16, 2021)