+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
 - Enh #138: Add "Save" action to sidebar navigation
 - Enh #9: Add activities for edited Wiki pages
 - Enh #140: Expand category hierarchy
+- Enh #113: Handle parallel editing
 
 
 1.5.3  (April 16, 2021)